+++ docs/changelog.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - enforce support of only drf-spectacular 0.27 and newer in pyproject.toml
 - ensure examples from `@extend_schema_serializer` are not ignored when adding error response examples
 - show default error response examples only when the corresponding status code is allowed
+- add `"null"` to the error code enum of `non_field_errors` validation errors
 
 ## [0.13.0] - 2024-02-28
 ### Changed