+++ docs/changelog.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [UNRELEASED]
+
+## [0.14.0] - 2024-06-19
 ### Added
 - declare support for DRF 3.15